# docs/swagger.yaml
basePath: /api
definitions:
  github_com_Dragodui_diploma-server_internal_modules_smarthome_homeassistant.HAState:
    properties:
      attributes:
        additionalProperties: true
        type: object
      entity_id:
        type: string
      last_changed:
        type: string
      last_updated:
        type: string
      state:
        type: string
    type: object
  models.AddDeviceRequest:
    properties:
      entity_id:
        type: string
      icon:
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
      room_id:
        type: integer
    required:
    - entity_id
    - name
    type: object
...
  models.ConnectHARequest:
    properties:
      token:
        minLength: 10
        type: string
      url:
        type: string
    required:
    - token
    - url
    type: object
  models.ControlDeviceRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      service:
        type: string
    required:
    - service
    type: object
...
  models.Home:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      invite_code:
        type: string
      memberships:
        description: relations
        items:
          $ref: '#/definitions/models.HomeMembership'
        type: array
      name:
        type: string
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
  models.HomeMembership:
    properties:
      created_at:
        type: string
      home:
        $ref: '#/definitions/models.Home'
      home_id:
        type: integer
      id:
        type: integer
      joined_at:
        type: string
      role:
        type: string
      status:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: relations
      user_id:
        type: integer
    type: object
...
  models.Room:
    properties:
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
  models.SmartDevice:
    properties:
      created_at:
        type: string
      entity_id:
        type: string
      home:
        $ref: '#/definitions/models.Home'
      home_id:
        type: integer
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      type:
        type: string
    type: object
...
  models.Task:
    properties:
      assignments:
        items:
          $ref: '#/definitions/models.TaskAssignment'
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      description:
        type: string
      due_date:
        type: string
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      reminder_minutes:
        type: integer
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      schedule:
        $ref: '#/definitions/models.TaskSchedule'
      schedule_type:
        type: string
    type: object
  models.TaskAssignment:
    properties:
      assigned_date:
        type: string
      complete_date:
        type: string
      created_at:
        type: string
      id:
        type: integer
      reminder_sent_at:
        type: string
      status:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: relations
      task_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TaskSchedule:
    properties:
      created_at:
        type: string
      current_rotation_index:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      next_run_date:
        type: string
      recurrence_type:
        description: daily, weekly, monthly
        type: string
      rotation_user_ids:
        description: JSON array of user IDs in order
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: relations
      task_id:
        type: integer
    type: object
...
  models.UpdateDeviceRequest:
    properties:
      icon:
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
      room_id:
        type: integer
    required:
    - name
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      memberships:
        description: relations
        items:
          $ref: '#/definitions/models.HomeMembership'
        type: array
      name:
        type: string
      profile_public:
        type: boolean
      task_assignments:
        items:
          $ref: '#/definitions/models.TaskAssignment'
        type: array
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Diploma project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Diploma Server API
  version: "1.0"
paths:
...
  /homes/{home_id}/rooms/{room_id}/devices:
    get:
      description: Get all devices assigned to a room
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SmartDevice'
            type: array
      security:
      - BearerAuth: []
      summary: Get devices by room
      tags:
      - smarthome
...
  /homes/{home_id}/smarthome/connect:
    post:
      consumes:
      - application/json
      description: Connect a Home Assistant server to the home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: HA Connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectHARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Connect Home Assistant
      tags:
      - smarthome
  /homes/{home_id}/smarthome/devices:
    get:
      description: Get all added devices for the home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SmartDevice'
            type: array
      security:
      - BearerAuth: []
      summary: Get devices
      tags:
      - smarthome
    post:
      consumes:
      - application/json
      description: Add a device from Home Assistant to the app
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Device Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add device
      tags:
      - smarthome
  /homes/{home_id}/smarthome/devices/{device_id}:
    delete:
      description: Remove device from the app
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete device
      tags:
      - smarthome
    get:
      description: Get device details with current state
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get device
      tags:
      - smarthome
    put:
      consumes:
      - application/json
      description: Update device name, room, or icon
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: integer
      - description: Device Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update device
      tags:
      - smarthome
  /homes/{home_id}/smarthome/devices/{device_id}/control:
    post:
      consumes:
      - application/json
      description: Call a service on the device (turn_on, turn_off, etc.)
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: integer
      - description: Control Command
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ControlDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Control device
      tags:
      - smarthome
  /homes/{home_id}/smarthome/disconnect:
    delete:
      description: Disconnect Home Assistant from the home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disconnect Home Assistant
      tags:
      - smarthome
  /homes/{home_id}/smarthome/discover:
    get:
      description: Discover all available devices from Home Assistant
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Dragodui_diploma-server_internal_modules_smarthome_homeassistant.HAState'
            type: array
      security:
      - BearerAuth: []
      summary: Discover devices
      tags:
      - smarthome
  /homes/{home_id}/smarthome/states:
    get:
      description: Get current states of all added devices
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Dragodui_diploma-server_internal_modules_smarthome_homeassistant.HAState'
            type: array
      security:
      - BearerAuth: []
      summary: Get all device states
      tags:
      - smarthome
  /homes/{home_id}/smarthome/status:
    get:
      description: Get connection status and config
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Home Assistant status
      tags:
      - smarthome
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"